+++ spec/shoulda/matchers/active_model/ensure_length_of_matcher_spec.rb
@@ -105,6 +105,52 @@ describe Shoulda::Matchers::ActiveModel::EnsureLengthOfMatcher do
     end
   end
 
+  context 'using translations' do
+    after { I18n.backend.reload! }
+
+    context "a too_long translation containing %{attribute}, %{model}" do
+      before do
+        stub_translation(
+          "activerecord.errors.messages.too_long",
+          "The %{attribute} of your %{model} is too long (maximum is %{count} characters)")
+      end
+
+      it "does not raise an exception" do
+        expect {
+          validating_length(:maximum => 4).should ensure_length_of(:attr).is_at_most(4)
+        }.to_not raise_exception(I18n::MissingInterpolationArgument)
+      end
+    end
+
+    context "a too_short translation containing %{attribute}, %{model}" do
+      before do
+        stub_translation(
+          "activerecord.errors.messages.too_short",
+          "The %{attribute} of your %{model} is too short (minimum is %{count} characters)")
+      end
+
+      it "does not raise an exception" do
+        expect {
+          validating_length(:minimum => 4).should ensure_length_of(:attr).is_at_least(4)
+        }.to_not raise_exception(I18n::MissingInterpolationArgument)
+      end
+    end
+
+    context "a wrong_length translation containing %{attribute}, %{model}" do
+      before do
+        stub_translation(
+          "activerecord.errors.messages.wrong_length",
+          "The %{attribute} of your %{model} is the wrong length (should be %{count} characters)")
+      end
+
+      it "does not raise an exception" do
+        expect {
+          validating_length(:is => 4).should ensure_length_of(:attr).is_equal_to(4)
+        }.to_not raise_exception(I18n::MissingInterpolationArgument)
+      end
+    end
+  end
+
   def validating_length(options = {})
     define_model(:example, :attr => :string) do
       validates_length_of :attr, options